+++ Realm/RLMObject.h
@@ -176,25 +176,23 @@
 /**
  Get objects matching the given predicate for this type from the default Realm.
  
- @param predicate   The argument can be an NSPredicate, a predicate string, or predicate format string
- which can accept variable arguments.
+ @param predicate   An NSPredicate to filter the array.
  
  @return    An RLMArray of objects of the subclass type in the default Realm that match the given predicate
  */
-+ (RLMArray *)objectsWhere:(id)predicate, ...;
++ (RLMArray *)objectsWhere:(NSPredicate *)predicate;
 
 /**
  Get an ordered RLMArray of objects matching the given predicate for this type from the default Realm.
  
- @param predicate   The argument can be an NSPredicate, a predicate string, or predicate format string
- which can accept variable arguments.
  @param order       This argument determines how the results are sorted. It can be an NSString containing
  the property name, or an NSSortDescriptor with the property name and order.
+ @param predicate   An NSPredicate to filter the array.
  
  @return    An RLMArray of objects of the subclass type in the default Realm that match the predicate
  ordered by the given order.
  */
-+ (RLMArray *)objectsOrderedBy:(id)order where:(id)predicate, ...;
++ (RLMArray *)objectsOrderedBy:(id)order where:(NSPredicate *)predicate;
 
 
 #pragma mark -